split mimo and nemotron tests by server start
